servidor/requirements.txt
flask
pillow
redis
requests
